HACK: disable GtkYieldMutex::Threads{Enter,Leave} recursion code
This breaks horribly when we encounter a nested event loop in Qt.
I still haven't figure out why exactly this works for GTK but doesn't
for the Qt case. But the issue is quite straightforward:

- GtkYieldMutex is acquired somewhere outside of ThreadsEnter
- a nested event loop is started up and ThreadsEnter gets called,
  yieldCounts is empty, acquire sets the count to 2.
- the nested event loop is exited, ThreadsLeave is called with the
  count set to 2. This gives n == 1 and release is going to be called
  twice, once for i == 0 and once for i == 1 (both are != n + 1, i.e. 2)

So essentially we unlocked the mutex from somewhere else, leading to
assertions when that lock guard tries to release the yield mutex.

I have not yet understood the idea behind ThreadsEnter/Leave, as I'd
personally have assumed that Enter/acquire would push to the
yieldCounts stack, and Leave would release and pop. But the code does
the push/pop in reverse for some reason.

So for now we just disable this and see what is going to break... At
least this makes the Qt file picker functional without asserting.

Change-Id: Ibe91c58e22fdd2de1783c8d93a3a0e991e583e6f