Use external process to launch KDE dialog
Using two XCB connections from the same process leads to problems.
E.g. when we want to access the clipboard contents from the GTK
LibreOffice application from within the KDE dialog, we would
run into basically a deadlock that times out after five seconds.
This is because the Qt XCB QPA does not run the eventloop when it
tries to synchronously read the clipboard contents.

There are probably more issues, which can be workarounded by using
a separate process, too.

This patch introdces the lo_kdefilepicker helper executable. It
launches a QFileDialog and talks with LO via a simple stdout/in IPC
protocol.

Change-Id: Ice87c4cbe5d152460760238f657ab8773670afa9