Do not send or read anything via IPC while we are waiting for response
When we wait for a response, the next write on the client side is
expected to answer this response. If we would now start writing other
commands, we could read the wrong response. And if we'd start reading
again, we'd have a race condition of multiple threads trying to read
concurrently.

This happens for a save file dialog, where the current filter is
tried to be read while we wait for the dialog to execute.

Change-Id: Iadce1727d72bdcac009537347006fcecf2a8cb31