Make the Gtk3/KDE5 file picker work for save dialogs
Fixup the handling of file name filters. Use the correct format that
Qt5 expects, i.e. "title (pattern)" and introduce a hash to correctly
map titles to and from the Qt5 filter strings.

Furthermore, ensure we set AnyFile as fileMode on the dialog, instead
of using ExistingFile(s).

This allows us to use the Gtk3/KDE5 file picker to save files (as).

Change-Id: I0d44583066b41b95de35c7fc3d4929d8d3d00d47