Initialize QApplication in the KDE5 backend when required
When we try to load the KDE5 backend into a vcl plug that does not
use Qt itself, e.g. for the Gtk3KDE5 backend, then no QApplication
is available. If so, initialize the QApplication to make the backend
functional.

In the future, we might want to introduce an out-of-process mechanism
here too. But for now this approach works as the backend does not
require event loop interactions which are error prone to get done.

Change-Id: Ifd7ce274735612fa4ad9de3dac1ebbfdb5d17a51